styled home page with login signup swap again
