+++ src/pages/StartupScreen/Startup.jsx
@@ -17,20 +17,32 @@ const Startup = ({user, setUser}) => {
       }, 2000);
     }
   };
-  
+
   const textRef = useRef(null);
+  const [isSigningUp, setIsSigningUp] = useState(false); // State to toggle between Login and Signup
+
+  const toggleAuthMode = () => {
+    setIsSigningUp((prevState) => !prevState); // Toggle between Login and Signup
+  };
 
   return (
-    <div>
-        <div className="auth-container" onClick={handleTextRotation}>
-          <div className="static-text" ref={textRef}>Databased</div>
-          <div>
-            <Login setUser={setUser} />
-            <Signup setUser={setUser} />
+      <div>
+        <div className="main-container" onClick={handleTextRotation}>
+          <div className="static-text" ref={textRef}>DataBaseD</div>
+          <div className="auth-container">
+            {isSigningUp ? (
+                <Signup setUser={setUser} />
+            ) : (
+                <Login setUser={setUser} />
+            )}
+            <button className={"swap-button"} onClick={toggleAuthMode}>
+              <span>{isSigningUp ? "Already have an account? Login" : "Need an account? Sign up"}</span>
+            </button>
           </div>
         </div>
-    </div>
-  );
-};
+      </div>
+  )};
+
+
 
 export default Startup;
